Updated MariaDB to 10.3 for CentOS 7.5

+++ centos/scripts/7/cleanup.sh
@@ -9,10 +9,9 @@ yum -y erase kernel-devel kernel-firmware kernel-headers gcc gcc-c++;
 rm -rf VBoxGuestAdditions_*.iso;
 
 # Clean up network interface persistence
-rm -f /etc/udev/rules.d/70-persistent-net.rules;
-mkdir -p /etc/udev/rules.d/70-persistent-net.rules;
-rm -f /lib/udev/rules.d/75-persistent-net-generator.rules;
-rm -rf /dev/.udev/;
+[ -f /etc/udev/rules.d/70-persistent-net.rules ] && rm -rf /etc/udev/rules.d/70-persistent-net.rules
+[ -f /lib/udev/rules.d/75-persistent-net-generator.rules ] && rm -rf /lib/udev/rules.d/75-persistent-net-generator.rules;
+[ -f /dev/.udev ] && rm -rf /dev/.udev/;
 
 for ndev in `ls -1 /etc/sysconfig/network-scripts/ifcfg-*`; do
     if [ "`basename $ndev`" != "ifcfg-lo" ]; then
